requirements.txt
aiogram==3.6.0
aiohttp
python-dotenv
aiosqlite
aiogram==3.13.0
requests==2.31.0
python-dotenv==1.0.0
sqlite3
